fix(react): simplify i18n context
