+++ packages/react/src/i18n/index.ts
@@ -17,13 +17,11 @@ export {
   registerI18n,
 } from '@videojs/core/i18n';
 
+export { I18nContext, useLocale, useTranslator } from './context';
 export type {
   CreateI18nOptions,
   CreateI18nResult,
   I18nContextValue,
   I18nProviderProps,
 } from './create-i18n';
-
-export { createI18n } from './create-i18n';
-export { I18nContext, useLocale, useTranslator } from './instance';
-export { I18nProvider } from './provider';
+export { createI18n, I18nProvider } from './create-i18n';
