--- packages/react/src/i18n/instance.tsx
@@ -1,7 +1,0 @@
-'use client';
-
-import { createI18nBase } from './base';
-
-export const i18nBase = createI18nBase();
-
-export const { I18nContext, useLocale, useTranslator } = i18nBase;
