+++ packages/react/src/player/container.tsx
@@ -3,7 +3,7 @@
 import type { HTMLAttributes, ReactNode, PointerEvent as ReactPointerEvent } from 'react';
 import { forwardRef, useEffect, useRef } from 'react';
 
-import { I18nProvider } from '../i18n/provider';
+import { I18nProvider } from '../i18n/create-i18n';
 import { useComposedRefs } from '../utils/use-composed-refs';
 import { useContainerAttach } from './context';
