+++ packages/react/src/presets/audio/minimal-skin.tsx
@@ -1,6 +1,6 @@
 import { cn } from '@videojs/utils/style';
 import { type ComponentProps, forwardRef, type ReactNode } from 'react';
-import { useTranslator } from '@/i18n/instance';
+import { useTranslator } from '@/i18n/context';
 import {
   CheckIcon,
   PauseIcon,
