+++ packages/react/src/presets/video/minimal-skin.tailwind.tsx
@@ -24,7 +24,7 @@ import {
 import { isString } from '@videojs/utils/predicate';
 import { cn } from '@videojs/utils/style';
 import { type ComponentProps, type CSSProperties, forwardRef, type ReactNode } from 'react';
-import { useTranslator } from '@/i18n/instance';
+import { useTranslator } from '@/i18n/context';
 import {
   AirPlayEnterIcon,
   AirPlayExitIcon,
